requirements.txt
pytest
pytest-cov
pytest-mock
gunicorn
flask
pyyaml
click
arrow
requests
psycopg2
voluptuous